+++ src/main/java/net/liukrast/repackaged/RepackagedConfig.java
@@ -15,6 +15,13 @@ public class RepackagedConfig {
     public static class Server {
         private Server() {}
         private static final ModConfigSpec.Builder BUILDER = new ModConfigSpec.Builder();
+        public static final ModConfigSpec.BooleanValue VANILLA_CRAFTER_UNPACKING = BUILDER
+                .comment("Makes so vanilla crafters respect order context from packaging requests")
+                .define("vanillaCrafterUnpacking", true);
+
+        public static final ModConfigSpec.BooleanValue BOX_ORDER_FIX = BUILDER
+                .comment("Makes so boxes always unpack in the correct order, even without passing through a re-packager")
+                .define("boxOrderFix", true);
 
         static final ModConfigSpec SPEC = BUILDER.build();
     }