+++ src/main/java/net/liukrast/repackaged/content/fluid/FluidTabScreen.java
@@ -3,9 +3,9 @@ package net.liukrast.repackaged.content.fluid;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.simibubi.create.content.logistics.BigItemStack;
 import com.simibubi.create.content.logistics.stockTicker.StockKeeperRequestMenu;
-import com.simibubi.create.content.logistics.stockTicker.StockTickerBlockEntity;
 import com.simibubi.create.foundation.gui.AllGuiTextures;
 import net.liukrast.deployer.lib.helper.GuiRenderingHelpers;
+import net.liukrast.deployer.lib.logistics.packager.screen.KeeperSourceContext;
 import net.liukrast.deployer.lib.logistics.packager.screen.StockTabScreen;
 import net.liukrast.repackaged.RepackagedConstants;
 import net.liukrast.repackaged.registry.RepackagedItems;
@@ -22,8 +22,8 @@ import java.util.Locale;
 
 public class FluidTabScreen extends StockTabScreen<Fluid, FluidStack> {
     private static final ResourceLocation TEXTURE = RepackagedConstants.id("textures/gui/fluid_stock_inventory.png");
-    public FluidTabScreen(StockTickerBlockEntity blockEntity, StockKeeperRequestMenu menu) {
-        super(blockEntity, menu, RepackagedItems.STANDARD_BOTTLES.getFirst().asItem(), RepackagedStockInventoryTypes.FLUID.get());
+    public FluidTabScreen(KeeperSourceContext context, StockKeeperRequestMenu menu) {
+        super(context, menu, RepackagedItems.STANDARD_BOTTLES.getFirst().asItem(), RepackagedStockInventoryTypes.FLUID.get());
     }
 
     @Override